No longer store iObjectRegistry in csRef<> as you're not supposed to do this
Using size_t to rather than (int) or (unsigned int) to iterate against STL
containers as it's more correct (though T::size_type is what is actually
returned by size() functions this normally equates to size_t)

No longer use csRef<>s as arguments to functions as it's not needed

Added support for "a3dl:clone" object.  To construct we take the template
object and use it's factory to make a new mesh.