iCamera: As reported in #1008, there was previously some problems around the management of the field of view and aspect ratio of the perspective cameras. The methods SetFOV() and SetFOVAngle() were working inconsistently, and it was not possible to achieve proper zoom effects. There was also a confusion whether the various parameters should be used in view space or in normalized screen coordinates.
Another related problem was the fact that, in a lot of different places within the CS code, it is assumed that the camera has a perspective projection, although this can be a custom one.

Still another related problem, although purely on the design point of view, was a duplication of the definition of the viewport dimensions. They were supposed to be defined both in the iView and in the iCamera, although that's the iView that is supposed to manage the definition of the viewport.

This patch and the following ones globally address all of these issues by introducing new FOV-related methods that are now supposed to work correctly, new camera projection methods that work for all camera types, and by concentrating all viewport-related methods within iView.

The main target achievement of this set of patches is to make both the resizing of the 2D canvas and the management of custom matrix cameras work much more correctly.

The list of changes for this iCamera patch is:
- Added the methods iPerspectiveCamera::Set/GetVerticalFOV(), Set/GetVerticalFOVAngle(), Set/GetAspectRatio() that are now supposed to work as intended.
- Deprecated all previous FOV-related methods. The aforementioned new methods should be used instead.
- Deprecated the method iCamera::SetViewportSize(). That's now the iView that manages the viewport dimensions.
- Added the generic methods iCamera::Project() and InvProject() that work in normalized screen coordinates and for all camera types.
- Deprecated the methods iCamera::Perspective() and InvPerspective(). Project() and InvProject() should be used instead.
- Warning: The methods GetShiftX/Y() and SetPerspectiveCenter() where previously inconsistently used both in screen space and in normalized screen coordinates depending on the camera interface being used. They are now always supposed to be provided in normalized screen coordinates.
- API documentation improvements.